extraction of utils-functions
