+++ src/App.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react'
 import Select from 'react-select'
-import { tail, zipObject, concat, sortBy, sumBy } from 'lodash'
 import NumberWithLabel from './components/NumberWithLabel'
 import HeaderText from './components/HeaderText'
+import getSum from './utils/getSum'
+import csvToJson from './utils/csvToJson'
 
 const axios = require('axios')
 
@@ -26,57 +27,15 @@ export default class App extends Component {
 
   componentDidMount = () => {
     axios.get('http://mockbin.org/bin/3f1037be-88f3-4e34-a8ec-d602779bf2d6').then(response =>
-      this.csvToJson(response.data)
+      this.setState(csvToJson(response.data, this.state.model))
     )
   }
 
   onChange = (value) => {
-    const sumClicks = this.getSum(value, 'clicks')
-    const sumImpressions = this.getSum(value, 'impressions')
+    const sumClicks = getSum(this.state.adwordData, value, 'clicks', this.state.model)
+    const sumImpressions = getSum(this.state.adwordData, value, 'impressions', this.state.model)
 
     this.setState({ value, sumClicks, sumImpressions })
-  }
-
-  getSum = (value, col) => {
-    const sum = sumBy(this.state.adwordData, (o) => {
-      if (o[this.getColumn(o, 'campaign')] === value || o[this.getColumn(o, 'channel')] === value) {
-        return o[this.getColumn(o, col)]
-      }
-
-      return 0
-    })
-
-    return sum
-  }
-
-  getColumn = (obj, prop) => Object.keys(obj)[this.state.model[prop]];
-
-
-  parseRow = (row) => {
-    const parsedRow = [...row]
-
-    parsedRow[2] = parseInt(parsedRow[2], 10)
-    parsedRow[3] = parseInt(parsedRow[3], 10)
-
-    return parsedRow
-  }
-
-  csvToJson = (csv) => {
-    const content = csv.trim().split('\n')
-
-    const header = content[0].split(',')
-    const adwordData = sortBy(tail(content).map(row =>
-      zipObject(header, this.parseRow(row.split(',')))
-    ), ['campaign', 'channel'])
-
-    const uniques = concat(
-      ...new Set(adwordData.map(item => item[this.getColumn(item, 'campaign')])),
-      ...new Set(adwordData.map(item => item[this.getColumn(item, 'channel')]))
-    )
-
-    const options = uniques.map(value => ({ value, label: value }))
-
-    this.setState({ options, adwordData })
   }
 
   render() {
